refactor(api): move metrics.js from monitoring to shared
